ADM-803[frontend]:compelete distinguish duplicated name in report page
